Add token & client helpers
Add some helpers for dealing with tokens, Tilapia & unauthenticated requests.
Make User.fetch_user use either the HTTP Basic password, or the API token,
whichever is available, and have the helper pass the token when there is one;
this allows it to call the API if the object has expired from Memecached.
A few more authn cleanups.