Re-organize routes & add Job view
Remove the "/organizations" controller, view & template (it was always
temporary).
Configure routes so that the organization name is always the first paramater.
Redirect directly to the first organization in the users membership list;
remove login_success route, view & template.
Add the Job template to retrieve (AJAX) and render job details.