Clean up job_info
Move the Javascript into its own external file & pull it in to the outer
templates where needed.
Namespace all of the "job info" elements & Javascript: use the "ji_" prefix
for anything related to displaying the job info.
Make the username & linkback URL global so that they can be used from the job
info Javascript without being inserted by Mustache: globals because a bunch of
functions are callbacks and it's far easier to pull data like that in via.
globals.