Add a user page
Add the controller, view, template & Javascript for the user page.
Add a jQuery MD5 plugin.
Move the User model under a Model namespace so we don't clash with the
Controller.
Change the navbar username to a link, to the users page.
Make the breadcrumbs more flexible; let the controller decide what is & isn't
a linkable element via. setting the 'url' attribute.
Redirect users to their user page if they are not a member of any
organization.