Make all of the search fields functional
Use the correct statuses for the list Status dropdown, and set appropriate
values for the status.
Parse the from & to fields into ISO dates.
Add the s_ prefix to the search parameter names.
Set gblOffset to the total number of records, which fixes initial chunk
loading where the total is less than the limit.