Add a helper binary
When the Gem is installed the statis assets will be in the folder where the
Gem was un-packed; the HTTP server needs to know the path to those assets in
order to properly server them. Add a simple executable that just prints the
absolute path to the static assets, so that they can be found.