Add a user-invaldation endpoint
Add /user/:username/invalidate as an endpoint that will flush the user object
from Memcache. This can be used to force a refresh when the users organization
membership changes.
When client-side authentication fails (E.g. an AJAX request gets 401) redirect
to / and let the server handle things. If the user session itself is still
valid (E.g. the organization was deleted by the user is still logged in) then
it can redirect somewhere useful; if the user really is not authenticated then
it can handle session invalidation and redirect to /login: this may also fix
the problem of users getting stuck on the /login page when login redirection
fails.